These are the files and an STM4Eclipse project to test the working of a comparator and ADC simultaneously.
This works on the STM32L05 Nucleo board, simply compile and flash.

It should be possible to run on ADC channel 1 / PA1 as well...
either I am doing something stupid, or I am an idiot.

Let me out of my misery if you find it!

Please see the caveats on Electronics.stackoverflow.com

http://electronics.stackexchange.com/questions/210292/stm32f0x1-comparator-comp-and-adc-on-same-pin-simultaneously/211418#211418

This is most useful since one can have an ultra fast hardware comparator as well as a direct reading of ADC as well as software window comparator all on one pin.

If the ST docs are interpreted fully, the fast comparator can be used to adjust a PWM and control e.g. current in a motor inverter.


ultra cool